Move currency related types into a new header
