+++ src/OpenLoco/openloco.vcxproj
@@ -202,6 +202,7 @@
     <ClInclude Include="Core\FileSystem.hpp" />
     <ClInclude Include="Core\Optional.hpp" />
     <ClInclude Include="Core\Span.hpp" />
+    <ClInclude Include="Currency.h" />
     <ClInclude Include="Date.h" />
     <ClInclude Include="Drawing\SoftwareDrawingEngine.h" />
     <ClInclude Include="Economy.h" />
